# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Casaura API
  version: 1.10.0-release-candidate
  description: Production-candidate contract for identity, tenancy, listings, marketplace, integrations, grounded AI, billing, privacy, and administration.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security: []
...
paths:
...
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      responses:
        "200": { description: Password reset and existing sessions revoked }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /auth/email/verify/{id}/{hash}:
    get:
      security: [{ sanctum: [] }]
      tags: [Authentication]
      operationId: verifyEmail
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: path, name: hash, required: true, schema: { type: string } }
      responses:
        "200": { description: Signed email address verified }
        "403": { description: Invalid or expired signature }
...
components:
  securitySchemes:
    sanctum:
      type: apiKey
      in: cookie
      name: laravel_session
...
  schemas:
...
    QualityCheck:
      type: object
      required: [key, complete, message]
      properties:
        key: { type: string, enum: [basics, location, description, features, media] }
        complete: { type: boolean }
        message: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: true }
            request_id: { type: string }
            current_version: { type: integer, minimum: 1 }
            checklist:
              type: array
              items: { $ref: "#/components/schemas/QualityCheck" }
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }